refactor: 타입 설정 폴더 및 파일 변경
